Fix test cases relying on undefined behavior
The way argmax handles ties changed between Pytorch 1.6 and 1.7, causing the
solution code to fail the run_model tests.  This commit adds a small asymmetry
to the output bias during model initialization to ensure a unique deterministic
argmax value.

Previously, the run_model test cases initialized all weights to the same value
and all biases to the same value, which meant that the model could output
identical values for its outputs when run by the student code (for example, on
the very first train batch for test_train_function and test_run_model, and all
batches in test_test_function).  In this case the argmax is undefined, and
versions of Pytorch prior to 1.7 weren't guaranteed to be consistent (and
returned different values than numpy argmax).